Changed background (first time gimp, don't hate me plox), added random stars

+++ galaxyconqueror/src/main/java/GalaxyConqueror/Model/Engine.java
@@ -10,13 +10,14 @@ import java.util.Random;
 import static GalaxyConqueror.Controller.ActionControl.*;
 import static GalaxyConqueror.Model.Constants.*;
 import static GalaxyConqueror.Model.Model.*;
-import static GalaxyConqueror.View.View.bullet;
-import static GalaxyConqueror.View.View.enemy;
+import static GalaxyConqueror.View.View.*;
+import static java.lang.Math.*;
 
 public class Engine {
     private static long enemySpawnLast = 0;
     private static long bulletDelayLast = 0;
     private static Random random=new Random();
+    private static int starShooters=0;
     public static void engine() {
         scoreLabel.setText("Score: " + score);
         hpLabel.setText("HP: " + player.hp);
@@ -30,6 +31,10 @@ public class Engine {
         if (enemySpawn >= ENEMY_SPAWN_TICK) {
             shootAndSpawnEnemies();
             enemySpawnLast = timeNow;
+        }
+        if(starShooters==0){
+            starShooters=random.nextInt(5)+5;
+            addStars(starShooters);
         }
 
         checkCollisions();
@@ -104,16 +109,18 @@ public class Engine {
 
     private static void checkCollisions () {
         for (Bullet b : bullets) {
-            //allied objects
+            //allied objects 0
             if (b.collisionId == 0) {
                 for (Enemy e : enemies) {
-                    if (b.me.getBoundsInParent().intersects(e.me.getBoundsInParent())) {
-                        e.subtractHp(b.getDmg());
-                        b.kill();
+                    if(e.collisionId==1) {
+                        if (b.me.getBoundsInParent().intersects(e.me.getBoundsInParent())) {
+                            e.subtractHp(b.getDmg());
+                            b.kill();
+                        }
                     }
                 }
             }
-            //enemy objects
+            //enemy objects 1
             if (b.collisionId == 1) {
                 if (b.me.getBoundsInParent().intersects(player.me.getBoundsInParent()) ) {
                     player.subtractHp(b.getDmg());
@@ -122,8 +129,10 @@ public class Engine {
             }
         }
         for (Enemy e : enemies) {
-            if (e.me.getBoundsInParent().intersects(player.me.getBoundsInParent())) {
-                player.kill();
+            if(e.collisionId==1) {
+                if (e.me.getBoundsInParent().intersects(player.me.getBoundsInParent())) {
+                    player.kill();
+                }
             }
         }
 
@@ -131,6 +140,32 @@ public class Engine {
 
     private static void gameOver () {
         System.out.println("Game over");
+    }
+
+    public static void addStars(int x){
+        Bullet starB=new Bullet(star,0,min(abs((random.nextDouble())+0.1),0.4),2);
+        for(int i=0;i<x;i++) {
+            Enemy starShooter = new Enemy(dot, 0, 1, 2) {
+                @Override
+                public void move() {
+                    double moveX=random.nextInt(50);
+                    if(x-moveX<0)
+                        x+=moveX;
+                    else if(x+moveX>SCREEN_WIDTH)
+                        x-=moveX;
+                    else{
+                        if(random.nextBoolean())
+                            x+=moveX;
+                        else
+                            x-=moveX;
+                    }
+                    me.relocate(x,y);
+                }
+            };
+            starShooter.addBullet(starB,random.nextInt(20)+2,random.nextInt(10)+3);
+            starShooter.setPosition(random.nextInt(SCREEN_WIDTH),0,90);
+            enemies.add(starShooter);
+        }
     }
 
 }
